🔥 (yes): Laying the foundation for a better light theme
First real commit of this project, aside the first one. This commit is mainly to fix some glaring issues, as well as lay the foundation for the full light theme support. I moved (almost) all the font import stuff from fonts.google.com to a TypeScript file to make the process better, I plan on doing the same for other repeting elements of the code but I currently  don't have the skill so...  Anyway this commit isn't that import so that's it!